Move supplemental Android test code to a dedicated module
