+++ androidtest/README.md
@@ -1,0 +1,4 @@
+# androidtest
+
+Includes helpers (arguably missing from `androidx.test:core`) for writing ["Local tests" and "Instrumented tests"](https://developer.android.com/training/testing/fundamentals#instrumented-vs-local).
+```
