+++ androidtest/src/main/java/org/odk/collect/androidtest/ActivityScenarioLauncherRule.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.testshared.android
+package org.odk.collect.androidtest
 
 import android.app.Activity
 import android.content.Intent
