+++ androidtest/src/main/java/org/odk/collect/androidtest/LiveDataTester.kt
@@ -1,15 +1,13 @@
-package org.odk.collect.testshared.android
+package org.odk.collect.androidtest
 
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.Observer
-import org.odk.collect.testshared.FakeLifecycleOwner
 
 class LiveDataTester {
 
     private val owner = FakeLifecycleOwner()
 
     fun <T> activate(liveData: LiveData<T>): LiveData<T> {
-        liveData.observe(owner, Observer { })
+        liveData.observe(owner) { }
         return liveData
     }
