+++ androidtest/src/main/java/org/odk/collect/androidtest/NestedScrollToAction.java
@@ -1,9 +1,11 @@
-package org.odk.collect.testshared.android;
+package org.odk.collect.androidtest;
 
+import static androidx.test.espresso.matcher.ViewMatchers.Visibility;
 import static androidx.test.espresso.matcher.ViewMatchers.isAssignableFrom;
 import static androidx.test.espresso.matcher.ViewMatchers.isDescendantOfA;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayingAtLeast;
 import static androidx.test.espresso.matcher.ViewMatchers.withEffectiveVisibility;
+import static androidx.test.espresso.util.HumanReadables.describe;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.anyOf;
 
@@ -17,8 +19,6 @@ import androidx.test.espresso.PerformException;
 import androidx.test.espresso.UiController;
 import androidx.test.espresso.ViewAction;
 import androidx.test.espresso.action.ScrollToAction;
-import androidx.test.espresso.matcher.ViewMatchers;
-import androidx.test.espresso.util.HumanReadables;
 
 import org.hamcrest.Matcher;
 
@@ -36,7 +36,7 @@ public class NestedScrollToAction implements ViewAction {
     @SuppressWarnings("unchecked")
     @Override
     public Matcher<View> getConstraints() {
-        return allOf(withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE), isDescendantOfA(anyOf(
+        return allOf(withEffectiveVisibility(Visibility.VISIBLE), isDescendantOfA(anyOf(
                 isAssignableFrom(ScrollView.class),
                 isAssignableFrom(HorizontalScrollView.class),
                 isAssignableFrom(NestedScrollView.class)
@@ -56,7 +56,7 @@ public class NestedScrollToAction implements ViewAction {
         if (!isDisplayingAtLeast(90).matches(view)) {
             throw new PerformException.Builder()
                     .withActionDescription(this.getDescription())
-                    .withViewDescription(HumanReadables.describe(view))
+                    .withViewDescription(describe(view))
                     .withCause(
                             new RuntimeException(
                                     "Scrolling to view was attempted, but the view is not displayed"))
