+++ androidtest/src/main/java/org/odk/collect/androidtest/NotificationDetails.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.testshared
+package org.odk.collect.androidtest
 
 import android.app.Notification
 import android.content.Intent
@@ -8,7 +8,7 @@ import org.robolectric.Shadows
  * Wraps a [Notification] object and provides access to details exposed by its Robolectric "Shadow"
  * without needing to explicitly depend on Robolectric.
  */
-class NotificationDetails(private val notification: Notification) {
+class NotificationDetails internal constructor(private val notification: Notification) {
 
     val contentText: String
         get() = Shadows.shadowOf(notification).contentText.toString()
