+++ androidtest/src/main/java/org/odk/collect/androidtest/ServiceScenario.kt
@@ -1,13 +1,11 @@
-package org.odk.collect.testshared.android
+package org.odk.collect.androidtest
 
 import android.app.NotificationManager
 import android.app.Service
 import android.content.Context
 import android.content.Intent
-import androidx.fragment.app.testing.FragmentScenario
 import androidx.lifecycle.Lifecycle
 import androidx.test.core.app.ActivityScenario
-import org.odk.collect.testshared.NotificationDetails
 import org.robolectric.Robolectric
 import org.robolectric.Shadows.shadowOf
 import org.robolectric.android.controller.ServiceController
