+++ collect_app/src/androidTest/java/org/odk/collect/android/support/rules/FormActivityTestRule.kt
@@ -12,9 +12,9 @@ import org.odk.collect.android.injection.DaggerUtils
 import org.odk.collect.android.storage.StorageSubdirectory
 import org.odk.collect.android.support.AdbFormLoadingUtils
 import org.odk.collect.android.support.pages.FormEntryPage
+import org.odk.collect.androidtest.ActivityScenarioLauncherRule
 import org.odk.collect.projects.Project
 import org.odk.collect.projects.Project.Companion.DEMO_PROJECT
-import org.odk.collect.testshared.android.ActivityScenarioLauncherRule
 import java.io.IOException
 
 class FormActivityTestRule @JvmOverloads constructor(
