+++ collect_app/src/test/java/org/odk/collect/android/activities/MainMenuActivityTest.kt
@@ -32,11 +32,11 @@ import org.odk.collect.android.support.CollectHelpers
 import org.odk.collect.android.utilities.ApplicationConstants
 import org.odk.collect.android.version.VersionInformation
 import org.odk.collect.androidshared.livedata.MutableNonNullLiveData
+import org.odk.collect.androidtest.ActivityScenarioLauncherRule
 import org.odk.collect.async.Scheduler
 import org.odk.collect.projects.Project
 import org.odk.collect.projects.ProjectsRepository
 import org.odk.collect.settings.SettingsProvider
-import org.odk.collect.testshared.android.ActivityScenarioLauncherRule
 
 @RunWith(AndroidJUnit4::class)
 class MainMenuActivityTest {
