+++ collect_app/src/test/java/org/odk/collect/android/activities/SplashScreenActivityTest.kt
@@ -23,9 +23,9 @@ import org.odk.collect.android.injection.config.AppDependencyModule
 import org.odk.collect.android.projects.CurrentProjectProvider
 import org.odk.collect.android.rules.MainCoroutineScopeRule
 import org.odk.collect.android.support.CollectHelpers
+import org.odk.collect.androidtest.ActivityScenarioLauncherRule
 import org.odk.collect.projects.ProjectsRepository
 import org.odk.collect.settings.SettingsProvider
-import org.odk.collect.testshared.android.ActivityScenarioLauncherRule
 
 @RunWith(AndroidJUnit4::class)
 class SplashScreenActivityTest {
