+++ collect_app/src/test/java/org/odk/collect/android/formentry/RecordingHandlerTest.java
@@ -1,6 +1,14 @@
 package org.odk.collect.android.formentry;
 
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.not;
+import static org.hamcrest.Matchers.nullValue;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule;
 
 import org.javarosa.core.model.data.StringData;
@@ -11,20 +19,12 @@ import org.junit.Test;
 import org.odk.collect.android.audio.AudioFileAppender;
 import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.widgets.support.FakeQuestionMediaManager;
+import org.odk.collect.androidtest.FakeLifecycleOwner;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.recording.RecordingSession;
-import org.odk.collect.testshared.FakeLifecycleOwner;
 
 import java.io.File;
 import java.util.HashSet;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.not;
-import static org.hamcrest.Matchers.nullValue;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 @SuppressWarnings("PMD.DoubleBraceInitialization")
 public class RecordingHandlerTest {
