+++ collect_app/src/test/java/org/odk/collect/android/widgets/utilities/AudioRecorderRecordingStatusHandlerTest.java
@@ -13,7 +13,7 @@ import org.odk.collect.android.formentry.FormEntryViewModel;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.recording.RecordingSession;
 import org.odk.collect.androidshared.livedata.MutableNonNullLiveData;
-import org.odk.collect.testshared.FakeLifecycleOwner;
+import org.odk.collect.androidtest.FakeLifecycleOwner;
 
 import java.util.function.Consumer;
