+++ geo/src/test/java/org/odk/collect/geo/LocationTrackerGeoPointViewModelTest.kt
@@ -10,11 +10,11 @@ import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import org.odk.collect.androidshared.livedata.MutableNonNullLiveData
+import org.odk.collect.androidtest.LiveDataTester
 import org.odk.collect.location.Location
 import org.odk.collect.location.satellites.SatelliteInfoClient
 import org.odk.collect.location.tracker.LocationTracker
 import org.odk.collect.testshared.FakeScheduler
-import org.odk.collect.testshared.android.LiveDataTester
 
 class LocationTrackerGeoPointViewModelTest {
