+++ permissions/src/test/java/org/odk/collect/permissions/PermissionDeniedDialogTest.kt
@@ -20,9 +20,9 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.odk.collect.androidtest.RecordedIntentsRule
 import org.odk.collect.fragmentstest.FragmentScenarioLauncherRule
 import org.odk.collect.testshared.RobolectricHelpers
-import org.odk.collect.testshared.android.RecordedIntentsRule
 
 @RunWith(AndroidJUnit4::class)
 class PermissionDeniedDialogTest {
