+++ settings.gradle
@@ -21,4 +21,5 @@ include ':upgrade'
 include ':permissions'
 include ':imageloader'
 include ':glide'
+include ':androidtest'
 include ':settings'
